Trailer Endpoint

+++ DECAMovies/src/main/java/com/example/DECAMovies/repository/FeignClientInterface.java
@@ -21,7 +21,7 @@ public interface FeignClientInterface {
     Optional<Object> singleMovie(@PathVariable("id") Integer id);
 
     @GetMapping(value = "/movie/{id}/videos", produces = "application/json")
-    Object trailers(@PathVariable Long id);
+    Object trailers(@PathVariable Integer id);
 
 }
